refactor: UserDto 클래스를 user feature 디렉토리로 이동
