+++ src/main/java/com/bluetoya/taradiddle/feature/auth/dto/SignInResponse.java
@@ -1,5 +1,7 @@
 package com.bluetoya.taradiddle.feature.auth.dto;
 
+import com.bluetoya.taradiddle.feature.user.UserDto;
+
 public record SignInResponse(AuthDto auth, UserDto user) {
 
 }
