+++ src/main/java/com/bluetoya/taradiddle/feature/user/UserController.java
@@ -1,9 +1,14 @@
 package com.bluetoya.taradiddle.feature.user;
 
 import com.bluetoya.taradiddle.common.ApiResponse;
-import com.bluetoya.taradiddle.feature.auth.dto.UserDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/users")
